backend/extract-doc-preview/requirements.txt
psycopg2-binary==2.9.9
boto3==1.34.0
requests==2.31.0
pdf2image==1.16.3
Pillow==10.2.0
